+++ ros2unbag/__init__.py
@@ -18,4 +18,4 @@ __all__ = [
     "TopicInfo",
 ]
 
-__version__ = "1.4.3"
+__version__ = "1.5.1"